requirements.txt
--find-links https://download.pytorch.org/whl/cu113/torch_stable.html
torch==1.11.0+cu113
torchvision==0.12.0+cu113
opencv-python
imgaug
lmdb
PyYAML
numpy
Levenshtein
Pillow